Further progress made
