+++ boa_engine/src/vm/opcode/return_stm/mod.rs
@@ -15,11 +15,11 @@ impl Operation for Return {
     const INSTRUCTION: &'static str = "INST - Return";
 
     fn execute(context: &mut Context<'_>) -> JsResult<ShouldExit> {
-        if let Some(finally_address) = context.vm.frame().catch.last().and_then(|c| c.finally) {
+        if let Some(finally_address) = context.vm.frame().try_catch.last().and_then(|c| c.finally()) {
             let frame = context.vm.frame_mut();
             frame.pc = finally_address as usize;
             frame.finally_return = FinallyReturn::Ok;
-            frame.catch.pop();
+            frame.try_catch.pop();
             let mut envs_to_pop = 0_usize;
             for _ in 1..context.vm.frame().env_stack.len() {
                 let env_entry = context
